Code clean and bug sweep: portfolio bumps, test fixes, dep cleanup
- main.py: remove redundant workers=1
- conventions.py: move import math to module level, remove redundant int() casts
- engine.py: align bump defaults with schema (0.001)
- schemas.py: add bump_spot_rel/vol_abs/rate_abs to LegInput for API consistency
- app.py: import ql_date_from_iso helper, use it in all 3 endpoints; remove
  redundant local sys import; pass portfolio leg bumps through to price_vanilla
- pyproject.toml & requirements.txt: remove unused python-dateutil
- test_logging.py: fix start_size byte-count bug (now tracks line count)
- test_api_contract.py: move import math to module level; add XML control-char
  sanitization test; add Accept-header edge-case test; use different strikes
  in portfolio aggregate math test
- test_properties.py: widen American tree tolerance to 2e-2

+++ src/desk_pricer/pricing/conventions.py
@@ -1,5 +1,6 @@
 """Day counts, calendars and convention helpers."""
 
+import math
 from datetime import date
 
 import QuantLib as ql
@@ -19,6 +20,5 @@ def default_day_count() -> ql.DayCounter:
 
 def expiry_from_t(valuation_date: ql.Date, t: float) -> ql.Date:
     # Round-half-up to avoid Python's banker's rounding bias
-    import math
-    days = max(1, int(math.floor(t * 365 + 0.5)))
-    return valuation_date + int(days)
+    days = max(1, math.floor(t * 365 + 0.5))
+    return valuation_date + days
